+++ packages/twenty-front/src/modules/auth/types/socialSsoSignInUp.type.ts
@@ -1,0 +1,4 @@
+export type SocialSsoSignInUpActionType =
+  | 'create-new-workspace'
+  | 'list-available-workspaces'
+  | 'join-workspace';