+++ packages/twenty-server/src/engine/core-modules/auth/strategies/oidc.auth.strategy.ts
@@ -11,7 +11,7 @@ import {
   AuthExceptionCode,
 } from 'src/engine/core-modules/auth/auth.exception';
 
-export type OIDCRequest = Omit<
+export type OidcRequest = Omit<
   Request,
   'user' | 'workspace' | 'workspaceMetadataVersion'
 > & {
@@ -26,7 +26,7 @@ export type OIDCRequest = Omit<
 };
 
 @Injectable()
-export class OIDCAuthStrategy extends PassportStrategy(
+export class OidcAuthStrategy extends PassportStrategy(
   Strategy,
   'openidconnect',
 ) {
@@ -87,7 +87,7 @@ export class OIDCAuthStrategy extends PassportStrategy(
     req: Request,
     tokenset: TokenSet,
     // oxlint-disable-next-line typescript/no-explicit-any
-    done: (err: any, user?: OIDCRequest['user']) => void,
+    done: (err: any, user?: OidcRequest['user']) => void,
   ) {
     try {
       const state = this.extractState(req);